[media] vimc: Virtual Media Controller core, capture and sensor
First version of the Virtual Media Controller.
Add a simple version of the core of the driver, the capture and
sensor nodes in the topology, generating a grey image in a hardcoded
format.

Signed-off-by: Helen Fornazier <helen.fornazier@gmail.com>

Changes since v1:
- Rename vmc to vimc
- Alignement adjustments
- Remove extra blank line in vimc Makefile
- Implement s_input in one line with return i ? -EINVAL : 0;
- Remove s_fmt_vid_cap and try_fmt_vid_cap, plug g_fmt_vid_cap directly
- Remove set_fmt  in vimc-sensor.c and plug get_fmt directly
- Reorder paramenters declaration in vimc_cap_start_streaming
- Fix typos irs-> isr, its->it's
- Remove obsolete comment
- Register the video device just before returning from vimc_cap_create
- Register the subdevice device just before returning from vimc_sen_create
- In vimc-sensor.c file, rename declared structure with prefix vimc_sd_
to vimc_sen_
- Remove vimc_cap_device/vimc_sen_device struct declaration from headers
- Remove vimc_cap_destroy/vimc_sen_destroy functions from headers and
declare them as static (they are already acessible by a pointer in the
struct returned in the create function)
- Rename vimc_pix_fmt_map{_list} to vimc_pix_map{_list}, shorter name
because it will have other maps used by the debayer for example that may
be called vimc_dep_pix_map{_list} and vimc_dep_pix_fmt_map{_list} is too
big
- Bug fix: check if stream is active in process frame function,
otherwise the frame may propagate to non active nodes
- Add node name in dev_dbg print