[media] vimc: Subdevices as modules
Change the core structure for adding subdevices in the topology.
Instead of calling the specific create function for each subdevice,
inject a child platform_device with the driver's name.
Each type of node in the topology (sensor, capture, debayer, scaler)
will register a platform_driver with the corresponding name through the
component subsystem.
Implementing a new subdevice type doesn't require vimc-core to be altered.

This facilitates future implementation of dynamic entities, where
hotpluging an entity in the topology is just a matter of
registering/unregistering a platform_device in the system.
It also facilitates other implementations of different nodes without
touching the core code and remove the need of a header file for each
type of node.

Signed-off-by: Helen Koike <helen.koike@collabora.com>

---

Series-changes: 5
[media] vimc: Subdevices as modules
	- Fix vimc_add_subdevs in rollback case. The loop variable can
	be negative, remove 'unsigned' from the loop variable and fix
	warning from smatch tool

Series-changes: 4
[media] vimc: Subdevices as modules
	- Rebase without [media] vimc: Optimize frame generation the through
	pipe
	- s/EXPORT_SYMBOL/EXPORT_SYMBOL_GPL
	- add struct vimc_platform_data to pass the entity's name to the
	sudmodule
	- Fix comment about vimc-input (remove vimc-output comment)

Series-changes: 3
[media] vimc: Subdevices as modules
	- This is a new patch in the series